+++ src/main/java/roomescape/common/exception/ExceptionController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.common.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
