+++ src/main/java/roomescape/view/PageController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
